requirements.txt
# Core Dependencies
pysc2==4.0.0
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2

# StarCraft 2 Integration
absl-py==1.4.0
sc2reader==1.8.0

# Machine Learning
gym==0.26.2
stable-baselines3==2.0.0
tensorboard==2.13.0

# Data Processing
pandas==2.0.3
h5py==3.9.0
pillow==10.0.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.0.74

# Utilities
tqdm==4.65.0
pyyaml==6.0
requests==2.31.0
scikit-learn==1.3.0

# Optional (for advanced features)
wandb==0.15.5  # For experiment tracking
ray==2.6.1  # For distributed training
